requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
ijson
requests
boto3
python-dotenv
